requirements.txt
pytest==8.4.1
requests==2.32.3
PyYAML==6.0.1
allure-pytest==2.14.0
faker==26.0.0